FE Test Dafidea
Make a Simple CRUD and AUTH with Vue.js and Tailwind CSS with a responsive design. This project includes authentication features such as login and logout, ensuring secure user access. The user data is displayed in a dynamic table, allowing users to view, add, edit, and delete entries seamlessly. The input and edit forms are fully integrated with the **GoRest GraphQL and REST API**, enabling real-time data management. The interface is styled with **Tailwind CSS** for a clean and modern look, while **Vue.js** ensures a smooth and interactive user experience.